Add prompt(), single_prompt(), and failsafe().
Severall functions have now been implemented. The prompt() and
single_prompt() are prompts that prompts the user. More information
will be on a new wiki page.

failsafe() is the failsafe mode function. This isn't finished yet, but
will be finished sooner or later. More information will be on a new
wiki page.