Update liblines.hpp
This commit brings the following changes:
  * Bump version number in about()
  * Add new parameter for show_copyright()

Because there is no longer the options screen, I've decided to replace
the options screen with a copyrigt message. show_copyright() will handle
this.

NOTE: Lines will still not display this message. This is because
main_screen() will be removed in the near future.

In about(), I did not update the version for Lines. That has now been
fixed.