Change directories to search for libraries
In the Makefile, it would search for liblines.so in the liblines
folder. However, liblines.so would not be located in there. The folder
no longer exists! (liblines/ exists in libs/).

In this commit, I've changed the directories to look in a build-libs
folder. This is a folder that exists.