Remove randomly pasted code
For some reason, there was some randomly pasted code in util.h. I
do not know how that ended up there, but I've removed that code.
The reason for removal is that it did not have proper synatx and was
duplicated from prompt() (and hence why I say 'randomly pasted code').