Change variable name in calc_coor_pts_slope()
Whenever you would compile an application or library with lines.h,
there will always be an error saying that slope couldn't be used as
a function. The main cause was how a variable was named 'slope' while
there was the conflicting liblines function 'slope()'. The variable
has been renamed to 'graph_slope' to prevent the naming conflict.