Update util.cpp
Many updates have been made to util.cpp. This includes the following:
  * Updates to the regular command line
  * Updates to the failsafe mode prompt
  * Added a copyright message

There have been various changes to the regular command line. The first
one being the addition of the "about" command. This features an updated
about screen along with a copyright message while featuring other things
as well.

The failsafe prompt is now a multi-prompt. The failsafe single-prompt will
be removed in upcomming changes. Other than that, only neccessary code
changes were done to make detecting commands compatible with the multi-prompt.

Before this commit, there was no copyright information in the source code. I've
finally updated this so that there is copyright information.