Various updates to util.h
Changes in util.h this time around include the following:

  * Add support for printing provided error information in failsafe mode
  * Failsafe mode now uses a multi-prompt

It has been decided that failsafe mode will now use a multi-prompt as though
there is no point to use a single-prompt. Eventualy, all single-prompts will
be phased out over time as though a multi-prompt as several advantages over
a single prompt.

Because failsafe mode would use a multi-prompt, I've changed the parameter
for the prompt type to print any custom error information. This would be
more useful as though it helps the user troubeshoot the issue that caused
Lines to enter failsafe mode.