Fix bug in prompt()
In prompt(), what is would do was prompt the user. It would do this
by saving the input into a variable and then it would return the a
variable (of type vector<string>). However, that variable would be
empty. What was supposed to happen was that the input would be split
up, but that didn't happen. This would cause a segmentation falt.

Now, thanks to Boost, this has been fixed. prompt() uses boost::split()
to split up the string and this prevents the segmentation falt error.