Add "gencoor" command and fix about command
I've added a new command called "gencoor". This command is meant
to replace options 2, 3, and 4. However, currently, gencoor has
options 2 and 3 implemented. Option 4 will soon be implemented.

gencoor (for now) has several options:
  * No option (default)
  * "write"

The default option (no option) represents option 2 while "write"
represents option 3. It is planned that there will be a "single"
option (representing option 4).

For the about command, previously, it did not return you to the
prompt like it was supposed to. This has since been fixed to return
you to main prompt.