Remove about and clear command
More and more functionality in option() (despite all of it being
commented out) is being removed with every commit. In this commit,
the about and clear commands have been removed. The clear command
won't see a revival and related functionality will be removed sooner
or later. The about command has already been added in liblines-util.