Remove deprecated functions
Several deprecated functions have been removed. They consist of the following:

  * get_platform()
  * show_copyright()
  * clear()
  * about()
  * help()

Most of these functions are for commands that options() would use. Most of
the commands that have relied on these functions have been either reimplemented
in liblines-util for will be reimplemented sooner or later. As for get_platform(),
it was unnecessary for that function and it would be better to use preprocessor
directives than this one function.

NOTE: The next deprecated function to be removed is main_screen(). It has been
around for some time but proves to be unnecessary and there are better ways of
having a main screen.