Removed calls of main_screen() and change messages
As part of an effort to remove anything that calls main_screen().
That has been deprecated and shouldn't be used anymore. Also, as
part of the process, I've also decided to change some messages as
well. This is for the main reason of trying to make Lines more
user friendly.