Improve algorithm in coordinate()
In coordinate(), for checking the delta X and delta Y to see if they
are both have a value of zero, that has been changed to see if only
has a value of zero. Also, coordinate() no longer checks the slope.
Because you can divide delta Y and delta X and get the slope, you
need to only check one because delta X or Y divided by 0 equals 0.
Because the slope is 0, we cannot find the y-intercept and equation.
So therefore coordinate() only checks delta X and delta Y to make
sure none of them equal 0.