Fix formatting of commands
In the copyright command, 2 new lines were created after it was
executed. This has been limited to 1 new line for cosmetic purposes
;). Also, when exit was executed, "Bye!" would not create a new
line and the formatting wouldn't look pretty. This has been fixed.
Finally, quotes have been added around an unknown command to better
help the user distinguish between the unknown command and the message,
and also for cosmetic reasons ;).