lines.hpp: remove copyright message supposed to be displayed on start
For whatever reason, the copyright message that was supposed to be
displayed when the program first starts is also displayed on the
about screen. This has been changed ad instead has a message that
tells the user to run the 'copyright' command for more information
about the copyright information on Lines.