lines.hpp: Update version and copyright (part 1), update help screen
1: Update version and copyright

To get prepared for the final release, the version and copyright have
been updated to show their correct versions. Now, the copyright date is
2019-2020, and the version is now marked as v3.0. Note that this only
affects copyright information in the about screen. It does not affect
the copyright date when Lines first starts up. However, that will be
fixed in the next commit

2: Update help screen

The help screen contained outdated information about gencoor. This has
since been updated to reflect the correct information about gencoor.