README.md: Fix formatting
There was a line under the 'Features' section which was on the third
bullet point. However, this line was supposed to be seperate and not
on the third bullet point. This has since been fixed so that the line
is on a seperate line and not on the third bullet point.