liblines.cpp: code cleanups, coordinate point generator changes
In this commit, I've updated the description of liblines.cpp in
the code, and also changed and added some comments as well.

The code for write_generated_points_to_file() has been removed.
It is still defined in liblines.hpp. Eventually, it will be re-
implemented...

As for the remaining coordinate point generators, they received
a very small update: one unnecessary variable was removed as it
proved unnecessary.