liblines.cpp: Improvements to finding equations, generating coordinate points
This commit brings several of the following changes:

  * Formatting improvements to slope()
  * One less variable has been removed from slope()
  * generate_point() now accepts any number of type double, not int.

slope() has seen some care too. Because the way it solves for an equation is
literally the same code behind coordinate() (with some changes), there is no
different. In this commit, slope() now has the same changes to coordinate():
one less variable while solving the equation and the formatting code (for
outputting the equation) has been greatly reduced.

generate_point() on the other hand didn't really change that much. It was
pretty limited and could only accept an integer. In this commit, that's been
changed. generate_point() now accepts any number that is a decimal. In other
words, the paramter for the number to use to generate the coordinate point
now accepts variables of type 'double'.