Makefile (Lines): Remove static liblines dependency
Previously, if you were to build the portable option of Lines, you
would need a static version of liblines. However, this has been
changed so that it compiles liblines directly into Lines without
doing any linking. This is more efficient in making the portable
version of Lines.

Also, just to mention, when making the portable version of Lines, the
makefile no longer puts the portable version of Lines in a seperate folder.
It will just be put in the root directory of the project folder.