liblines Makefile: rename 'portable' option to 'static'
Ever since the last commit, it has become irrelevant to keep the
portable option, but may still be relevant to keep the portable
option for purposes of building a static version of liblines. In
this commit, I have renamed the 'portable' option to 'static'.

Along with the rename, a few things have changed: 1) the liblines
Makefile no longer makes a directory for the portable version of Lines
and 2) the Makefile no longer moves it to that directory.