Update to reflect latest events
The README needed to have an update to reflect the latest events in Lines. Now, you can only generate a coordinate point 2 different ways.

I also tweaks the link to the CONTRIBUTING.md file. It now points to that file locally rather than using a web link. Also, finally, I separated the note at the top of the page saying that Lines isn't a graphing calculator.