lines.hpp: display new version string if 'DEBUG' is present, added fraction support, added include guards
Here, thanks to the new 'debug' target in the Makefile, if the
'DEBUG' macro is defined, then we display a new version string
(note that the version string is current incorrect, but that'll
change in the beta release of v3.1).

As promised, fraction support has finally come to Lines. However,
fraction support is quite wonky, and if there are spaces present,
then Lines will display an error. This will be fixed later on.

The include guards in this file have been because why not? Let's
make this a formal C++ header, shall we? :)