frac.hpp: Better exception handling, imrpoved fraction handling
Exceptions, wherever they may be thrown, are now caught, regardless
of the exception (or at least, almost all of them, if that's the case).
However, the result is almost always 0, and this will be changed later.

Rather than erasing whitespaces arond the separator, all whitespaces
are erased wherever any are found.