Added support for doxygen to generate documentation
I've always been trying to document Lines in a good fashion. However, I've
never been able to do so. As I continue to learn things, I've finally
discovered Doxygen. This way, I can document the internals of Lines and finally
make contributing easier and actually possible (because now people can see how
the internals of Lines works very easily).

However, to address the issue of the wiki, this will not replace the wiki in its
entirety. This only replaces the wiki for internal documentation of Lines (although,
yes, it never really was a thing), and any "external" parts of Lines (such as options,
basics, etc) will still be documented on the wiki. In fact, documentation with Doxygen
is intended to be a supplement to the overall documentation of Lines, not a replacement.