liblines: removed unnecessary headers, removed write_generated_coordinate_points_to_file()
This version of liblines finally comes with the removal of
write_generated_coordinate_points_to_file(). This was a declared and
as of v3.0 unused function. Finally, this has been removed.

I also gave liblines a bit of love. I removed all the unnecessary headers
there were and added include guards in liblines.hpp.