Restructure a few things in Lines
A few things were restructured in Lines to make things work
better. This will all be in Lines v3.1 beta 1 (v3.1b1). Below
are some highlights of the restructuring:

  * A new system of option handling and such
  * Removal of Boost as a dependency
  * Complete fraction AND proper coordinate point support
  * frac.hpp has been renamed to lcl_interpret.hpp (and with
    the appropriate lcl_interpret.cpp as well).

These should hopefully help make Lines better in various ways
than before, and hopefully this makes a better experience in
maintaining Lines. More will be detailed on the v3.1b1 release
page on GitHub.

However, there's also 2 new files (among the other new files
introduced in the restructuring): lcl/value.hpp and lcl/value.cpp.
These files are not yet used in Lines, but they will be used in
future versions of Lines. These are to add a bit of evaluation support
on the Lines Command Line (LCL). However, these files are unfinished
and thus can be ignored for now.

And finally, Happy New Year to anyone who's reading this!