Make symbols with extern C
Symbols have now been made with C linkage, making liblines suitable
for being loaded at runtime.

However, previous versions of Lines (binary-wise) may require that
liblines be built with mangled symbols in order for them to work.
This is no longer the case, and so building a regular version of
liblines will fail because the required symbols are technically
missing (or different). Therefore, a new make target has been made
that compiles liblines with mangled symbols, which is called
'compat'. This builds liblines with mangled symbols which should make
liblines work with older versions of Lines.